Change drop amount to int and CraftedDrink to CraftedIngredient for shaker

+++ Assets/Sofia/scripts/Ingredients/Alcohol/AlcoholSO.cs
@@ -17,6 +17,7 @@ public class AlcoholSO : ScriptableObject
 
     public AlcoholType alcoholType;
     public Material dropColor;
-    public float dropValue;
+    public int dropValue;
+  
  
 }
